feat(main): Add Linode and DigitalOcean fetchers
Adds Linode and DigitalOcean to the list of IP range
fetchers. Includes corresponding updates to README.md and
generated.go.  Also adds a TorFetcher implementation,
though it is disabled by default. It converts
individual Tor exit nodes to /32 and /128 CIDR ranges.
Explains the default disabling of Tor and ASN fetchers
in the README.  Improves formatting in main README.